+++ alacritty_terminal/src/term/mod.rs
@@ -1423,21 +1423,19 @@ impl<T: EventListener> Handler for Term<T> {
                 if self.mode.contains(TermMode::ALT_SCREEN) {
                     self.grid.reset_region(..);
                 } else {
-                    // Change the absolute vi cursor position into inside the viewport, but
-                    // preserve the relative one in the display area.
-                    let line = self.vi_mode_cursor.point.line + self.grid.display_offset();
-                    self.vi_mode_cursor.point.line = line.grid_clamp(self, Boundary::Cursor);
-
                     self.grid.clear_viewport();
+
+                    self.vi_mode_cursor.point.line =
+                        self.vi_mode_cursor.point.line.grid_clamp(self, Boundary::Cursor);
                 }
 
                 self.selection = None;
             },
             ansi::ClearMode::Saved if self.history_size() > 0 => {
-                let line = self.vi_mode_cursor.point.line + self.grid.display_offset();
-                self.vi_mode_cursor.point.line = line.grid_clamp(self, Boundary::Cursor);
-
                 self.grid.clear_history();
+
+                self.vi_mode_cursor.point.line =
+                    self.vi_mode_cursor.point.line.grid_clamp(self, Boundary::Cursor);
 
                 self.selection = self.selection.take().filter(|s| !s.intersects_range(..Line(0)));
             },
@@ -2127,7 +2125,7 @@ mod tests {
     }
 
     #[test]
-    fn clear_viewport_updates_vi_mode_cursor_pos() {
+    fn clear_viewport_set_vi_cursor_into_viewport() {
         let size = SizeInfo::new(10.0, 20.0, 1.0, 1.0, 0.0, 0.0, false);
         let mut term = Term::new(&Config::default(), size, ());
 
@@ -2136,53 +2134,37 @@ mod tests {
             term.newline();
         }
 
-        // Emulate vi cursor motions on default configurations.
-        // gg
+        // Change the display area and the vi cursor position.
         term.scroll_display(Scroll::Top);
-        term.vi_mode_cursor = term.vi_mode_cursor.scroll(&term, 10);
-        // jjjjj
-        for _ in 0..5 {
-            term.vi_mode_cursor = term.vi_mode_cursor.motion(&mut term, ViMotion::Down);
-        }
-        // lll
-        for _ in 0..3 {
-            term.vi_mode_cursor = term.vi_mode_cursor.motion(&mut term, ViMotion::Right);
-        }
+        term.vi_mode_cursor.point = Point::new(Line(-5), Column(3));
 
         // Clear the viewport.
         term.clear_screen(ansi::ClearMode::All);
 
-        assert_eq!(term.vi_mode_cursor.point.line.0, 5);
+        assert_eq!(term.grid.display_offset(), 0);
+        assert_eq!(term.vi_mode_cursor.point.line.0, 0);
         assert_eq!(term.vi_mode_cursor.point.column.0, 3);
     }
 
     #[test]
-    fn clear_scrollback_updates_dispaly_area() {
+    fn clear_scrollback_set_vi_cursor_into_viewport() {
         let size = SizeInfo::new(10.0, 20.0, 1.0, 1.0, 0.0, 0.0, false);
-        let mut term = Term::new(&MockConfig::default(), size, ());
+        let mut term = Term::new(&Config::default(), size, ());
 
         // Create 10 lines of scrollback.
         for _ in 0..29 {
             term.newline();
         }
 
-        // Emulate vi cursor motions on default configurations.
-        // gg
+        // Change the display area and the vi cursor position.
         term.scroll_display(Scroll::Top);
-        term.vi_mode_cursor = term.vi_mode_cursor.scroll(&term, 10);
-        // jjjjj
-        for _ in 0..5 {
-            term.vi_mode_cursor = term.vi_mode_cursor.motion(&mut term, ViMotion::Down);
-        }
-        // lll
-        for _ in 0..3 {
-            term.vi_mode_cursor = term.vi_mode_cursor.motion(&mut term, ViMotion::Right);
-        }
+        term.vi_mode_cursor.point = Point::new(Line(-5), Column(3));
 
         // Clear the scrollback buffer.
         term.clear_screen(ansi::ClearMode::Saved);
 
-        assert_eq!(term.vi_mode_cursor.point.line.0, 5);
+        assert_eq!(term.grid.display_offset(), 0);
+        assert_eq!(term.vi_mode_cursor.point.line.0, 0);
         assert_eq!(term.vi_mode_cursor.point.column.0, 3);
     }